🔥 remove: home 폴더 삭제, HomePage -\> LoginPage로 변경
